Moved @tryghost/expres-dynamic-redirects to services lib folder
fixes https://linear.app/ghost/issue/ENG-2120/express-dynamic-redirects

- this commit moves the lib code and tests to the core folder so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
- I'm not entirely happy about the location of this, but it's shared
  between 2 services
- at the very least, I'd rather get the file moved over to the core
  folder and then we can rejiggle things around afterwards
