+++ ghost/core/package.json
@@ -83,7 +83,6 @@
     "@tryghost/email-service": "0.0.0",
     "@tryghost/email-suppression-list": "0.0.0",
     "@tryghost/errors": "1.3.5",
-    "@tryghost/express-dynamic-redirects": "0.0.0",
     "@tryghost/ghost": "0.0.0",
     "@tryghost/helpers": "1.1.90",
     "@tryghost/html-to-plaintext": "0.0.0",
