--- ghost/express-dynamic-redirects/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/DynamicRedirectManager');
