--- ghost/express-dynamic-redirects/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@tryghost/express-dynamic-redirects",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura --check-coverage -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "should": "13.2.3"
-  },
-  "peerDependencies": {
-    "express": "^4.17.1"
-  }
-}
